Runtime/Dvm: Use uint64_t instead of size_t for the number of intersecting bytes because size_t is only 32bits wide on 32bit hosts and the value can exceed the range
svn:sync-xref-src-repo-rev: r76288

+++ src/VBox/Runtime/common/dvm/dvm.cpp
@@ -1,4 +1,4 @@
-/* $Id: dvm.cpp 40027 2012-02-07 23:09:31Z alexander.eichner@oracle.com $ */
+/* $Id: dvm.cpp 40137 2012-02-15 16:24:51Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT Disk Volume Management API (DVM) - generic code.
  */
@@ -482,7 +482,7 @@ RTDECL(int) RTDvmMapQueryBlockStatus(RTDVM hVolMgr, uint64_t off, uint64_t cb,
     {
         PRTDVMVOLUMEINTERNAL pVol;
         bool fVolFound = false;
-        size_t cbIntersect;
+        uint64_t cbIntersect;
         uint64_t offVol;
 
         /*
